Move SqliteConnectionExtensions to Shared Kernel
