+++ application/account-management/Tests/Api/Signups/SignupTests.cs
@@ -4,6 +4,7 @@ using FluentAssertions;
 using NSubstitute;
 using PlatformPlatform.AccountManagement.Core.Database;
 using PlatformPlatform.AccountManagement.Core.Signups.Commands;
+using PlatformPlatform.SharedKernel.Tests.Persistence;
 using PlatformPlatform.SharedKernel.Validation;
 using Xunit;
